Add a multistage version of Dockerfile usable in autobuilds
This change adds a multistage build of the operator binary and
operator image that does not need build args. As such, it's
usable in autobuilds by systems like quay.io. Symlinks give
it multiple command targets, so the same image is usable in
multiple systems (kftcl vs opendatahub-operator for example).