Add "none" component profile (#284)
* Add "none" component profile

Also set it as the default, to better enable explicit component selection.

Enforce default via CRD schema definition, rather than controller code, to ensure that the implicit "none" propogates to DataScienceClusters after creation as a transformation enforced at the API.

* Update CSV description with "none" profile

Additionally, rework the description in the CSV base so that the formatting is easier to work with in the future and help reduce accidental misalignments, or broken markdown syntax.