deploy: DeployManifestsFromPath: pass Component when appropriate
There is bunch of issues (tracker [1]) when it can be useful to
amend subcomponent's manifests based on the DataScienceCluster
component configuration and to do it in similar way like namespace
and labels are added in DeployManifestsFromPath.

Extend the API to have access to the configuration and pass the
structure. In some case the function is used not for the DSC's
formal subcomponent, like monitorig. Pass nil for that case.

[1] https://github.com/opendatahub-io/opendatahub-operator/issues/441

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>