components, deploy: implement deployment/spec/replica overriding
Implement kustomize patching Component's Deployment's .spec.replicas
based on the DataSciencCluster configuration.

Adds Replicas field to the Component to have easy access to the
manifest data as a field.  Uses *int for `json:,omitempty` to work
properly (do not put 0 by default). Extends ComponentInterface to
have access to the field.

Note: the path in the patch must start from /,
otherwise (spec/replicas) it adds it to the root (.replicas) and
complains that it is not in the scheme (Deployment object).

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>