tests: e2e: install servicemesh opertators before test run
Introduce setUp() method for testContext to install dependency operators,
- install servicemeshoperator and serverless-operator
- enable servicemesh in DSCI
- check that DSCI and DSC reflect the configuration change.

Since it is supposed that user install them (and user removes) it is
not implemented as additional tests but setUp. Probably would be
fine to have proper uninstallation in tearDown in case the test
installed them but removing of Subscriptions and CSVs is not
enough (checked, got broken setup) and implementation of proper
cleanup does not make sense at the moment.

To work with client.Client methods add
github.com/operator-framework/api/pkg/operators/v1alpha1 types to
the scheme.

Update Kserve check, it does not need special case (error
"Please install the operator before enabling component") handling
since its dependencies are installed.

DSCI/DSC validation code checks only servicemesh related parts
mostly for simplicity. But there is no any more special
configuration there to get unexpected values.

Spec values in setupSubscription() which creates an
instance of Subscription structure are basically predefined, the
operators are provided by redhat-operators catalog, must be
installed into openshift-operators namespace and so
on. InstallPlanApproval was added to workaround CI problems (see
below) but did not help alone.

In the CI the way it installs opendatahub-operator makes its and
following Subscriptions manually approved. As the result creating
Subscription objects for the operators does not cause operators
installation, it hungs waiting for InstallPlan approvement. To fix
it the code finds the InstallPlans and patches (server-apply
patching is used) it to get approved. It requires Force option since
the fields belong to 'catalog'. ClusterServiceVersionNames is copied
from the original object otherwise server complains about missing
field. It also required Kind and APIVersion.

ensureServicemeshOperators() takes *testing.T just for
logging. Channel is used instead of waitGroup to collect errors from
every operator installation which are done in parallel to speed up
the process.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>