fix(manifests): return walkErr immediately (#842)
`walkErr` is propagated as part of a recursive call when traversing a directory tree.

This adds a guard check and does not proceed if the error happens in the call chain.

Without this, the controller will `panic` in case of the wrong path defined for a manifest.

(cherry picked from commit 052124e32ae9488d252ccc804aab8dcfe910bdf7)