tests: e2e: rewrite with helpers
Move routine functionality to a set of helpers and make the
testcases short enough to focus on the actual actions and checks.

The naming follows testify package naming plus own conventions:

- require* -- fails immediately when not sutisfied
- assert*  -- makes the test failing but continues, returns bool
- require/assertEventually -- runs polling internally
- ensure*  -- uses existing resource or creates a new one
- set/getObject* -- operations on the Object (Component, Deployment)

It does not use testify's *.Eventually due to lack of
flexibility. Implemented version sets 'immediate' true and able to
stop polling when error occurs.

The functionality of reusing existing DSC and DSCI instances
is preserved.

Change polling interval to 2s to speed up. With correct checks of
preconditions it works (before it relied on the time)

Replace waitings with polling for condition.

Reorganize the suites, it allows for example to run validation tests
in parallel. With correct environment setup it's possible now to run
the tests individually.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>