fix: aligns provider name between template and cleanup logic
This is short term fix for existing codebase. Long term we need to
improve the way of how we are storing config information to limit cases
where we rely on pre/suffixes. Cases like this should be kept as it's
own thing instead, as it represents the concept in the
infrastructure/authz setup.