feat(tpl): fails when key does not exist
The templates used in Feature SDK are not intended (at least right now) to provide conditional rendering like in Helm. Thus we should fail fast if key referred in a template is not provided in the Feature.

This is achieved by passing parsing option (missingkey=error) to built-in Golang templating.