feat(tmpl): fails when key does not exist (#916)
The templates used in Feature SDK are not intended (at least right now) to provide conditional rendering like in Helm. Thus we should fail fast if the Feature's underlying map does not provide the key referred to in a template. 

This is achieved by passing the parsing option (missingkey=error) to built-in Golang templating.

> [!NOTE]
> This works out of the box with `struct`s but we should also fail when dealing with `map`s.