feat(build): adds gci to go fmt target
This update incorporates the gci tool into the make fmt command, aiming to improve the formatting of Go files and ensure that imports adhere to our established conventions.

Unfortunately, due to how .golangci.yml is defined, we are not able to just use --disable-all --enable=gci flags. This yields an error as we use enable-all and disable in the config file for linter.

To address this, the golangci-lint configuration is dynamically adjusted using the yq tool, which is now included in our toolchain.

This process creates a temporary configuration file specifically for golangci-lint execution, ensuring gci's integration without altering the primary linting setup.

Furthermore, the generated temporary file is added to the CLEANFILES list, ensuring its removal when calling make clean, maintaining a clean build environment.

To prevent this temporary file from being tracked by Git, a *.mktmp.* pattern has been added to our .gitignore, safeguarding against accidental inclusion of temporary files in the repository.