plugins: add RemoverPlugin, field remover
Make it the same way is other builtin plugins are done implementing
an RNode filter.

Since it is supposed to be called actually to a single resource, not
the whole ResourceMap, implement extra method TransformResource.

Use Pipe*() instead of direct filter calling even that it calls only
one filter.

- first time to call the methods of the inhereted RNode object of
the Resource;
- second (PipeE) to automatically skip the result

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>