deploy: pass resource.Resource to manageResource
Do not convert Resource to Unstructured early but just before final
changes in Unstructured before operations on the cluster.

This will allow to run kustomize filters after getResource() call,
meaning after analysing resource state in the cluster.

This is supposed to be used to filter out whitelisted fields from
the manifests if the resource already exist and leave them as
default values if the resource is being created by the operator.

Unfortunately, it makes a bit of code duplication since the Resource
to Unstructured conversion have to be done for both create and patch
branches.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>