deploy: skipUpdateOnWhitelistedFields: convert to RemoverPlugin
Use more universal way to filter whitelisted fields.

Make it possible to configure list of the fields as an array of
configured RemoverPlugins and define components to remove in a map
"Component Name" -> "pointer to the list of the fields". It makes it
possible to define them separately and reuse the list of the
fields (array of RemoverPlugins) for different components.

skipUpdateOnWhitelistedFields now takes Resource, so make the
conversion to Unstructured after it.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>