cluster_config: make GetClusterServiceVersion return NotFound error (#1028)
GetClusterServiceVersion() may return nil error and nil
pointer. On one hand it's ok to check object for nil in not found
condition but it's a bit counter intuitive to have err nil when
there is acutally "not found"

It caused a problem in GetVersion where this nil condition was not
checked.

Refactor it to return the error and handle it properly.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>