build: add manager target to Makefile and use it in Dockerfile
Hide the actual building logic in the Makefile and call it from
Dockerfile. It will allow to add build flags without changing
Dockerfiles.

Add additional target to not run all the dependencies of "bulid" and
output to the root directory what Dockerfiles usually expect.

Keep environment variables in the Dockerfile since they may change
for different usecases (multiarch build for example).

The Makefile is not copied by default, add that.

The same change should be done for other Dockerfiles.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>