Compile in build time version
Version is basically hardcoded (into Makefile and several other
places) so it's possible to compile it in and provide it then
instead of dummy 0.0.0 one.

Add version package and update manager target in the Makefile to set
the version during build time with `-ldflags -X`.

The value must be a string so use semver.Parse() to save it into
semver.Version.

Use recursively expanded variable (=) to allow possible version
manipulations.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>