Makefile: fetch GOLANG_VERSION from go.mod
A bit of automation for golang version update.

For local builds make it possible to set it only once in go.mod (it
actually means minimal required version, but in the project it is
used as the version for builds) and then reuse it for Dockerfiles:
- pass it as the build argument for normal image build (Dockerfile);
- use it for TOOLBOX_GOLANG_VERSION for toolbox.Dockerfile.

For compatibility to keep workflow of overriding IMAGE_BUILD_FLAGS
--build-arg USE_LOCAL working introduce another variable
IMAGE_BUILD_FLAGS_ALL to combine GOLANG_VERSION build argument with
the passed IMAGE_BUILD_FLAGS.

Add target golang-version-update to update the default version in
the Dockerfiles to avoid doing it manually in all of the places and
just commit then for the case of using Dockerfile without Makefile.

This change does not affect integration, version their should still
be updated separatly.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>