cluster: cache GetPlatform() value
Jira: https://issues.redhat.com/browse/RHOAIENG-8483

Platform is discovered on start up in main.go and is not supposed to
change during program lifetime but it can be time consuming since it
performs many api calls.

The patch caches the platform value and returns it if it's set.

It uses nil pointer for the blank value. There is no concurency
access problems since when it's stored once from main.go all
accesses is read-only.

The patch potentially changes behaviour since now it returns nil
error when the actual platform lookup would possible make a error
but the actual error is caught in main.go on the first lookup.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>