Dsc creation postpone (#1041)
* main: move DSC creation after DSCIs one

Reorder a bit to keep definition of cleanExistingResourceFunc next
to the Add() call.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>

* main: convert DSC creation to RunableFunc

DSC creation is checked by the webhook so will not work from the
main before manager starts if webhook is enabled.

The same as DSCI. It also requires CreateWithRetry() for the same
reasons as

e26100e87e53 ("upgrade: retry if default DSCI creation fails (#1008)")

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>

---------

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>