chore: reworks authorino istio injection
Instead of performing patching of Authorino deployment as part of
`PostConditions` hook, it is now a `Feature` on its own.

As a result, we no longer need the `ApplyManifest` mehtod for the `Feature` struct.

This function was created solely to apply a single manifest as an `Action` and
was used only for this specific use case. With the dedicated feature, a deployment
patch can now be defined as a regular manifest source and included as part of the Apply phase.